requirement.txt
itsdangerous  ==  2.1.2
Jinja2 == 3.1.2
MarkupSafe   ==  2.1.1
mysql-connector-python  == 8.0.31
pip      ==   22.3.1
protobuf   ==  3.20.1
setuptools   ==  65.5.0
Werkzeug     == 2.2.2
flask == 2.2.2
